+++ vpin-studio-commons/src/main/java/de/mephisto/vpin/commons/fx/ConfirmationDialogController.java
@@ -1,5 +1,6 @@
 package de.mephisto.vpin.commons.fx;
 
+import javafx.application.Platform;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ButtonType;
@@ -77,6 +78,10 @@ public class ConfirmationDialogController implements DialogController {
     if(okText != null) {
       okButton.setText(okText);
     }
+
+    Platform.runLater(() -> {
+      okButton.requestFocus();
+    });
   }
 
   public Optional<ButtonType> getResult() {